+++ torchbearer/callbacks/checkpointers.py
@@ -181,18 +181,23 @@ class Interval(_Checkpointer):
         save_model_params_only (bool): If `save_model_params_only=True`, only model parameters will be saved so that
             the results can be loaded into a PyTorch nn.Module. The other option, `save_model_params_only=False`,
             should be used only if the results will be loaded into a Torchbearer Trial object later.
+        period (int): Interval (number of steps) between checkpoints
+        on_batch (bool): If true step each batch, if false step each epoch.
         period (int): Interval (number of epochs) between checkpoints
         pickle_module (module): The pickle module to use, default is 'torch.serialization.pickle'
         pickle_protocol (int): The pickle protocol to use, default is 'torch.serialization.DEFAULT_PROTOCOL'
     """
 
-    def __init__(self, filepath='model.{epoch:02d}-{val_loss:.2f}.pt', save_model_params_only=False, period=1,
-                 pickle_module=torch.serialization.pickle, pickle_protocol=torch.serialization.DEFAULT_PROTOCOL):
+    def __init__(self, filepath='model.{epoch:02d}-{val_loss:.2f}.pt', save_model_params_only=False, period=1, on_batch=False, pickle_module=torch.serialization.pickle, pickle_protocol=torch.serialization.DEFAULT_PROTOCOL):
 
         super(Interval, self).__init__(filepath, save_model_params_only=save_model_params_only,
                                        pickle_module=pickle_module, pickle_protocol=pickle_protocol)
         self.period = period
         self.epochs_since_last_save = 0
+
+        if on_batch:
+            self.on_step_training = self.on_checkpoint
+            self.on_checkpoint = lambda _: None
 
     def state_dict(self):
         state_dict = super(Interval, self).state_dict()